Split Tile.hpp into Tile.hpp and TileTCount.hpp
