+++ include/RAJA/pattern/kernel.hpp
@@ -180,6 +180,7 @@ RAJA_INLINE void kernel(SegmentTuple &&segments, Bodies &&... bodies)
 #include "RAJA/pattern/kernel/Lambda.hpp"
 #include "RAJA/pattern/kernel/ShmemWindow.hpp"
 #include "RAJA/pattern/kernel/Tile.hpp"
+#include "RAJA/pattern/kernel/TileTCount.hpp"
 
 
 #endif /* RAJA_pattern_kernel_HPP */
